Added in all the operation functions. Error handling is at the moment from my perspective pretty good. Whats left is the big part. An algorithm that actually sorts it in as few moves as possible

+++ initial_try/libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: akuburas <akuburas@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/26 07:40:02 by akuburas          #+#    #+#             */
-/*   Updated: 2024/02/14 06:37:34 by akuburas         ###   ########.fr       */
+/*   Updated: 2024/02/16 04:39:18 by akuburas         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ typedef struct s_list
 	void			*content;
 	struct s_list	*next;
 }					t_list;
-int		ft_atoi(const char *str);
+long	ft_atoi(const char *str);
 void	ft_bzero(void *s, size_t n);
 void	*ft_calloc(size_t count, size_t size);
 int		ft_isalnum(int c);
